requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic[email]==2.8.2
pydantic-settings==2.3.4  # <-- ADD THIS LINE
python-dotenv==1.0.1
sqlalchemy==2.0.31
alembic==1.13.2
livekit-api>=1.0.0